requirements.txt
Flask==3.1.0
torch==2.2.0
transformers==4.49.0
faiss-cpu==1.10.0
groq==0.18.0
python-dotenv==1.0.1
numpy==1.26.4
